fix the order of initialization

+++ WS08/Rectangle.cpp
@@ -16,7 +16,7 @@ namespace sdds {
 	Rectangle::Rectangle() :m_width(0), m_height(0) {}
 
 	Rectangle::Rectangle(const char* label, int width, int height) :LblShape(label), m_width(width), m_height(height) {
-		if (m_height < 3 || m_width < (strlen(label) + 2)) {
+		if (m_height < 3 || m_width < static_cast<int>(strlen(label) + 2)) {
 			m_height = 0;
 			m_width = 0;
 		}
